feat(user-ui): remove project_box
